Rewrite connection handling, timeouts, config.
Basically, everything got rewritten.

Ping handling is now done properly, so there should be no
possible way for the bot to kill itself improperly now.

The startup code now returns a promise, and checks to see
if the server likes what it sent. Error messages are checked
and the bot quits if it cannot continue.

Config values have changed now that capabilities is actually
handled, multiple nicknames can be passed, and the irc-socket
wraps a net-socket.

The license was changed from BSD to ISC. They're legally
equivalent.