Refactoring, Rewrite Timeout, More Startup Tests
The timeout stuff is cross-cutting, so they should be in scope for
everything, and now they are. It also now does what the server does,
and sends a PING waiting for a PONG. It defaults to taking ten
minutes to determining a full timeout, and there's no config to
change that currently.

The onData handler was documented for what it was doing.

The onLine handler was simplified a lot. Now it's just back to
detecting PINGs and sending PONGs.

I made the output for the MockSocket for readable, so I can see what
is going on a lot more esaier now.

All the nickname cases are now well tested.